Minor refactors: tweak task package name and make Task abstract
