+++ src/main/java/command/Command.java
@@ -1,6 +1,6 @@
 package command;
 
-import tasks.TaskList;
+import task.TaskList;
 import ui.Ui;
 
 public abstract class Command {
