+++ src/main/java/task/EventTask.java
@@ -1,4 +1,4 @@
-package tasks;
+package task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
